Fixed compilation errors

+++ src/Player.h
@@ -2,7 +2,7 @@
 #ifndef PLAYER_H
 #define PLAYER_H
 #include "INC_SDL.h"
-#include "physics.hpp"
+#include "physics.h"
 #define MAX_SPEED 50
 
 
@@ -15,7 +15,7 @@ class Player
 
 
 		//Set the position of the player on screen
-		void setPosition(int x, int y);
+		void setPosition(double x, double y);
 
 			
 		//Methods that can be called from model class
